requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
bert-extractive-summarizer==0.10.1
blinker==1.9.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.10
defusedxml==0.7.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
flask==3.1.0
flask-wtf==1.2.2
fonttools==4.55.3
fpdf2==2.8.2
fsspec==2024.12.0
gunicorn==23.0.0
huggingface-hub==0.27.1
idna==3.10
itsdangerous==2.2.0
jieba3k==0.35.1
jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language-data==1.3.0
lxml==5.3.0
lxml-html-clean==0.4.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.11
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.1
packaging==24.2
pdfkit==1.0.0
pillow==11.1.0
pip==24.2
preshed==3.0.9
pydantic==2.10.4
pydantic-core==2.27.2
pygments==2.19.1
pyparsing==3.2.1
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
safetensors==0.5.1
scikit-learn==1.6.0
scipy==1.15.0
setuptools==75.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
sympy==1.13.1
thinc==8.3.3
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing-extensions==4.12.2
urllib3==2.3.0
validators==0.34.0
wasabi==1.1.3
weasel==0.4.1
werkzeug==3.1.3
wheel==0.44.0
wordcloud==1.9.4
wrapt==1.17.0
wtforms==3.2.1
